glue-airflow-local/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glue-airflow-local"
version = "0.1.0"
description = "Translate AWS Glue Workflow Terraform definitions into Airflow DAGs that run locally with mocks."
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Haris Bin Saif" }]
requires-python = ">=3.10"
dependencies = [
    "python-hcl2>=4.3.0",
    "apache-airflow>=2.9,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
glue-airflow-local = "glue_airflow_local.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/glue_airflow_local"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP", "RUF"]

[tool.mypy]
strict = true
# Pin to the lowest supported runtime so 3.11+ syntax that would break 3.10 users is caught.
python_version = "3.10"
files = ["src"]

[[tool.mypy.overrides]]
module = ["hcl2", "hcl2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
